Update build-react-android.sh merkle

+++ scripts/build-react-android.sh
@@ -22,30 +22,12 @@ if ! command -v cmake &> /dev/null; then
   exit 1
 fi
 
-n_cpu=1
-if uname -a | grep -q "Darwin"; then
-  n_cpu=$(sysctl -n hw.logicalcpu)
-elif uname -a | grep -q "Linux"; then
-  n_cpu=$(nproc)
+
+if ! command -v cmake &> /dev/null; then
+  echo "cmake found"
+  exit 1
 fi
 
-t0=$(date +%s)
-
-cd "$ROOT_DIR/android/src/main"
-
-# Build the Android library (arm64-v8a)
-cmake -DCMAKE_TOOLCHAIN_FILE="$CMAKE_TOOLCHAIN_FILE" \
-  -DANDROID_ABI=arm64-v8a \
-  -DANDROID_PLATFORM="$ANDROID_PLATFORM" \
-  -DCMAKE_BUILD_TYPE="$CMAKE_BUILD_TYPE" \
-  -B build-arm64
-
-cmake --build build-arm64 --config Release -j "$n_cpu"
-
-mkdir -p jniLibs/arm64-v8a
-
-# Copy the library to the jniLibs folder
-cp build-arm64/*.so jniLibs/arm64-v8a/
 
 rm -rf build-arm64
